requirements.txt
aiogram==3.22.0
fastapi==0.118.0
httpx==0.28.1
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pypdf==6.1.1
uvicorn[standard]==0.37.0
